Added edit option to viewnote.
Default is to view notes with no ability to edit them. In this case, the
notes are viewed using chromium's markdown renderer (needs Markdown
reader chrome extension installed.)

Had to change directory structure to accomodate this. Files generated
for viewing are stored in PREFIX/search, and the notes themselves are
stored in PREFIX/notes. Contents of search are deleted whenever there
more than 20 files in there.